Bundled convertDate function into Answer component
